backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
qdrant-client==1.7.0
python-dotenv==1.0.0
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
asyncio==3.4.3
pydantic==2.5.0
python-multipart==0.0.6